Add script to emulate NI System Web Server
This script will emulate the reboot web service usually provided by the NI
System Web Server.  Note that the reboot web service will restart the LabVIEW
run-time daemon and will not actually reboot the system.  Also note that the
server script does not perform any authentication so it should be used with
caution.