Fix power off issue with labview service
If the target is not shut down properly (power is abruptly turned off),
the labview run-time daemon would not start properly on a subsequent boot.
This happened because a schroot session file, that is usually deleted when a
session is ended properly, was left behind.

Change the labview.service to call a script which checks for the existence
of the session file and deletes it before starting the labview daemon.