NISysServer: Fix synchronous os.system() call
The RebootEx web service actually restarts the LabVIEW run-time daemon, but
since the HTTP handler doesn't send the response until after the user's
handler returns, this can cause the response to be delayed long enough that
the host thinks that an error has occurred.

To avoid this situation, spawn a daemon thread to restart the LabVIEW service
while allowing the handler to return a response in parallel.