Hard link host resolv.conf to chroot
By default resolv.conf is copied to a chroot when a session is started.  This
can lead to the host system's resolv.conf becoming out of sync with the
chroot, and then name resolution within the chroot can fail.

Instead hard link the host's resolv.conf to the chroot to ensure that any
updates will be synced in both file systems.