Fix LV project restart issue
When the LV runtime is restarted from the LV project, from the Restart
command or from a startup application install, LabVIEW calls into SysCfg
to perform the restart.  SysCfg expects that an IP address is passed to
the restart API, NOT a hostname.  The SysCfg restart call in turn
contacts the NISysServer.py script to get the target's IP address and then to
perform the actual restart operation.

In certain scenarios, the target board may have a network connection and
no internet connection, and in this case getIP() would return something
like 'beaglebone.localdomain'.  Change getIP() to use 'hostname -I' to
get the IP without requiring an internet connection.  If all else fails,
also return an IP address rather than a fully qualified hostname.