make_lvrt_image: Add RPi hardware config
For Raspberry Pi images, run the raspi-config script to enable various
hardware that is used by the LINX IO API.

In order to make this work, mount the boot partition from the RPi image
since raspi-config needs to touch some files here (config.txt).

Note that running raspi-config in the offline case will result in some
errors that can be ignored, since it calls modprobe and access
device-tree locations that are not accessible in the container.